feat: uses gid as process id in trisolaris

+++ server/controller/db/metadb/model/platform_rsc_model.go
@@ -76,7 +76,7 @@ type Process struct {
 	Name           string    `gorm:"column:name;type:varchar(256);default:''" json:"NAME" mapstructure:"NAME"`
 	VTapID         uint32    `gorm:"column:vtap_id;type:int;not null;default:0" json:"VTAP_ID" mapstructure:"VTAP_ID"`
 	PID            uint64    `gorm:"column:pid;type:int;not null;default:0" json:"PID" mapstructure:"PID"`
-	GID            uint64    `gorm:"column:gid;type:int;not null;default:0" json:"GID" mapstructure:"GID"`
+	GID            uint32    `gorm:"column:gid;type:int;not null;default:0" json:"GID" mapstructure:"GID"`
 	DeviceType     int       `gorm:"column:devicetype;type:int;default:null" json:"DEVICE_TYPE" mapstructure:"RESOURCE_TYPE"`
 	DeviceID       int       `gorm:"column:deviceid;type:int;default:null" json:"DEVICE_ID" mapstructure:"RESOURCE_ID"`
 	PodGroupID     int       `gorm:"column:pod_group_id;type:int;default:null" json:"POD_GROUP_ID" mapstructure:"POD_GROUP_ID"`
